refactor: init refactor of pkg files
