+++ core-service/pkg/controller/error.go
@@ -1,4 +1,4 @@
-package http
+package controller
 
 import "fmt"
